Start of staff service
